アラーム設定画面のUI変更
アラーム設定画面のUIを変更
